Update Project endpoints for new version
The new version adds pagination support and a new way of returning data
based on this pagination support. The pagination settings are passed as
query parameters and we receive the Projects as an array inside the
response object. Both the WorkspaceProject and PersonalProject are
combined into one as I felt there is not a need to have them as separate
since they have so many similar keys. This also avoids having to add
more logic to identify the types and have them as different types in one
array.