Refactor pagination response and options
This now makes it easier to use the same pagination struct for different
endpoints by using generics. It helped with the collaborator endpoint
and will help when implementing the tasks endpoint. I have also updated
the PaginationOptions struct to PaginationFilters to better reflect its
purpose and match closer to how I have made the request body be called
options and query params be called filters.