Make addQueryParams work with non-string values
This ensures we can pass non-string values in the query parameters. For
example we can pass a boolean value into this query params while still
keeping the proper type in our go code. This fixes the issue when we
pass `Limit` as a query parameter to the `GetComments` function as it is
an integer in the API.