Move sync api models to pkg/todoist
I have moved the sync api to be held also under pkg/todoist. This is to
make it so that I can reuse the same models for both the sync api and
the rest api. This is because the sync api is now the same as the rest
api and some models are shared between the two.