Add better error responses for API calls
Previously we were comparing if the response status code was not equal
to what was expected and then we would return the status code back as an
error but without any message from the Todoist server. This fixes that
problem by also returning the response body in the error when we make
our `APIError.Error()` gets called.

Todoist has also refactored their API so calls that previously return
204 now return 200. Making it easier to just check if we have an error.